requirements.txt
fastapi==0.123.4
uvicorn==0.38.0
sqlmodel==0.0.27
python-dotenv==1.2.1
itsdangerous==2.2.0
jinja2==3.1.6
python-multipart==0.0.20
passlib==1.7.4
bcrypt==4.2.1
psycopg2-binary==2.9.10
